# openapi.yaml
openapi: 3.0.3
info:
  title: Kitchen Helper
  version: 1.0.0
  description: Kitchen Helper
paths:
...
  /api/recipes/recipes/:
    post:
      operationId: api_recipes_recipes_create
      description: Create a new recipe. Requires authentication.
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
        required: true
      security:
      - jwtAuth: []
      - basicAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRecipe'
          description: ''
    get:
      operationId: api_recipes_recipes_list
      description: Retrieve a list of recipes. Admins can filter by draft or published
        status. Regular users can only see published recipes.
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: status
        schema:
          type: string
        description: Filter recipes by status (published -- for usual users, draft|all
          -- for staff)
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCreateRecipeList'
          description: ''
  /api/recipes/recipes/{id}/:
    patch:
      operationId: api_recipes_recipes_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Recipes.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCreateRecipeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCreateRecipeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCreateRecipeRequest'
          '*/*':
            schema:
              $ref: '#/components/schemas/PatchedCreateRecipeRequest'
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRecipe'
          description: ''
    get:
      operationId: api_recipes_recipes_retrieve
      description: Retrieve a specific recipe by ID.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Recipes.
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRecipe'
          description: ''
    put:
      operationId: api_recipes_recipes_update
      description: Update an existing recipe. Requires authentication.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Recipes.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
        required: true
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRecipe'
          description: ''
    delete:
      operationId: api_recipes_recipes_destroy
      description: Delete a specific recipe by ID. Requires authentication.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Recipes.
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '204':
          description: No response body
  /api/recipes/favorite_recipes/:
    post:
      operationId: api_recipes_favorite_recipes_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteRecipeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/FavoriteRecipeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FavoriteRecipeRequest'
          '*/*':
            schema:
              $ref: '#/components/schemas/FavoriteRecipeRequest'
        required: true
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteRecipe'
          description: ''
    get:
      operationId: api_recipes_favorite_recipes_list
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FavoriteRecipe'
          description: ''
  /api/recipes/favorite_recipes/{id}/:
    delete:
      operationId: api_recipes_favorite_recipes_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Recipes.
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '204':
          description: No response body
    get:
      operationId: api_recipes_favorite_recipes_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Recipes.
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteRecipe'
          description: ''
...
  /api/food/base-ingredients/{id}/:
    patch:
      operationId: api_food_base_ingredients_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Base ingredient.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedBaseIngredientRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedBaseIngredientRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedBaseIngredientRequest'
          '*/*':
            schema:
              $ref: '#/components/schemas/PatchedBaseIngredientRequest'
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseIngredient'
          description: ''
    get:
      operationId: api_food_base_ingredients_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Base ingredient.
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseIngredient'
          description: ''
    put:
      operationId: api_food_base_ingredients_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Base ingredient.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseIngredientRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BaseIngredientRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BaseIngredientRequest'
          '*/*':
            schema:
              $ref: '#/components/schemas/BaseIngredientRequest'
        required: true
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseIngredient'
          description: ''
    delete:
      operationId: api_food_base_ingredients_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this Base ingredient.
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '204':
          description: No response body
...
  /api/food/user-ingredients/{id}/:
    patch:
      operationId: api_food_user_ingredients_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this User ingredient.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedUserIngredientRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedUserIngredientRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedUserIngredientRequest'
          '*/*':
            schema:
              $ref: '#/components/schemas/PatchedUserIngredientRequest'
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserIngredient'
          description: ''
    get:
      operationId: api_food_user_ingredients_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this User ingredient.
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserIngredient'
          description: ''
    put:
      operationId: api_food_user_ingredients_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this User ingredient.
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIngredientRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserIngredientRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserIngredientRequest'
          '*/*':
            schema:
              $ref: '#/components/schemas/UserIngredientRequest'
        required: true
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserIngredient'
          description: ''
    delete:
      operationId: api_food_user_ingredients_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this User ingredient.
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - basicAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
    BaseIngredient:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        description:
          type: string
        photo:
          type: string
          format: uri
          nullable: true
        slug:
          type: string
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        cat:
          type: integer
      required:
      - cat
      - id
      - name
      - slug
    BaseIngredientRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        photo:
          type: string
          format: binary
          nullable: true
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        cat:
          type: integer
      required:
      - cat
      - name
      - slug
...
    CreateRecipe:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        slug:
          type: string
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        photo:
          type: string
          format: uri
          nullable: true
        content:
          type: string
        instructions:
          type: string
        cat:
          type: integer
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        author:
          type: integer
          readOnly: true
          nullable: true
        is_published:
          allOf:
          - $ref: '#/components/schemas/IsPublishedEnum'
          title: Pub/Draft
        equipments:
          type: array
          items:
            $ref: '#/components/schemas/REquipment'
      required:
      - author
      - cat
      - slug
      - title
    CreateRecipeRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        photo:
          type: string
          format: binary
          nullable: true
        content:
          type: string
        instructions:
          type: string
        cat:
          type: integer
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientRequest'
        is_published:
          allOf:
          - $ref: '#/components/schemas/IsPublishedEnum'
          title: Pub/Draft
        equipments:
          type: array
          items:
            $ref: '#/components/schemas/REquipmentRequest'
      required:
      - cat
      - slug
      - title
...
    FavoriteRecipe:
      type: object
      properties:
        email:
          type: string
        slug:
          type: string
      required:
      - email
      - slug
    FavoriteRecipeRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
      required:
      - email
      - slug
...
    Ingredient:
      type: object
      properties:
        base_ingredient_name:
          type: string
        user_ingredient_name:
          type: string
        quantity:
          type: integer
        unit:
          type: string
      required:
      - quantity
      - unit
    IngredientRequest:
      type: object
      properties:
        base_ingredient_name:
          type: string
        user_ingredient_name:
          type: string
        quantity:
          type: integer
        unit:
          type: string
          minLength: 1
      required:
      - quantity
      - unit
...
    IsPublishedEnum:
      enum:
      - false
      - true
      type: boolean
      description: |-
        * `False` - DRAFT
        * `True` - Published
...
    PaginatedCreateRecipeList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/CreateRecipe'
    PatchedBaseIngredientRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        photo:
          type: string
          format: binary
          nullable: true
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        cat:
          type: integer
...
    PatchedCreateRecipeRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        photo:
          type: string
          format: binary
          nullable: true
        content:
          type: string
        instructions:
          type: string
        cat:
          type: integer
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientRequest'
        is_published:
          allOf:
          - $ref: '#/components/schemas/IsPublishedEnum'
          title: Pub/Draft
        equipments:
          type: array
          items:
            $ref: '#/components/schemas/REquipmentRequest'
...
    PatchedUserIngredientRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        user:
          type: integer
...
    REquipment:
      type: object
      properties:
        equipment_name:
          type: string
        quantity:
          type: integer
      required:
      - equipment_name
      - quantity
    REquipmentRequest:
      type: object
      properties:
        equipment_name:
          type: string
          minLength: 1
        quantity:
          type: integer
      required:
      - equipment_name
      - quantity
...
    UserIngredient:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        description:
          type: string
        user:
          type: integer
      required:
      - id
      - name
      - user
    UserIngredientRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
        user:
          type: integer
      required:
      - name
      - user
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT